requirements.txt
Flask-Cors==3.0.3
Flask-Migrate==2.1.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Marshmallow==2.13.6
psycopg2==2.7.3.1
pycrypto==2.6.1
SQLAlchemy==1.1.13
